requirements.txt
tensorflow-gpu==2.2.0
keras==2.3.1
numpy==1.18.4
matplotlib